fix(ghdl): Moving GHDLsimulator and parser into testing folder
